imported UTF-8 functions
BitTorrent/1.0 tells that all strings must be encoded as UTF-8, regardless
the encoding of the system of the end user of the library (be it something
totally different), it should be able to encode strings as UTF-8, so we
need these routines.

The strings coming from the user's system will be decoded by the functions
provided in wchar.h.

This means we support reading and writing UTF-8 and reading the user's OS's
encoding.  For other encodings, external tools may be used to convert the
strings.