Make key search recursive by default
This makes searching keys easier within a bencoded structure.
One doesn't have to know the data to search a key, as the function
will now search it within ALL the dictionaries of the bencoded data.

Only the first occurence of each key can be retrieved, but this should
not be necessary as all keys are unique within a torrent, except for
what's under the "files" key. One would have to iterate bsearchkey()
calls through each member of the "files" list.

I still think it's worth it!